+++ src/Form/IntakeForm.php
@@ -514,16 +514,30 @@ class IntakeForm extends FormBase {
       '#required' => TRUE,
     ];
 
-    // Grazing acreage.
-    $form['land_use']['grazing_acreage'] = [
+    // Rangeland acreage.
+    $form['land_use']['rangeland_acreage'] = [
       '#type' => 'number',
-      '#title' => $this->t('Grazing acreage'),
+      '#title' => $this->t('Rangeland acreage'),
       '#min' => 0,
       '#step' => 0.1,
-      '#default_value' => $saved_values['land_use']['grazing_acreage'] ?? '',
+      '#default_value' => $saved_values['land_use']['rangeland_acreage'] ?? '',
       '#states' => [
         'visible' => [
-          ':input[name="property[land_use][land_use][grazing]"]' => ['checked' => TRUE],
+          ':input[name="property[land_use][land_use][rangeland]"]' => ['checked' => TRUE],
+        ],
+      ],
+    ];
+
+    // Pasture acreage.
+    $form['land_use']['pasture_acreage'] = [
+      '#type' => 'number',
+      '#title' => $this->t('Pasture acreage'),
+      '#min' => 0,
+      '#step' => 0.1,
+      '#default_value' => $saved_values['land_use']['pasture_acreage'] ?? '',
+      '#states' => [
+        'visible' => [
+          ':input[name="property[land_use][land_use][pasture]"]' => ['checked' => TRUE],
         ],
       ],
     ];
@@ -566,6 +580,20 @@ class IntakeForm extends FormBase {
       '#states' => [
         'visible' => [
           ':input[name="property[land_use][land_use][rowcrops]"]' => ['checked' => TRUE],
+        ],
+      ],
+    ];
+
+    // Forestry acreage.
+    $form['land_use']['forestry_acreage'] = [
+      '#type' => 'number',
+      '#title' => $this->t('Forestry acreage'),
+      '#min' => 0,
+      '#step' => 0.1,
+      '#default_value' => $saved_values['land_use']['forestry_acreage'] ?? '',
+      '#states' => [
+        'visible' => [
+          ':input[name="property[land_use][land_use][forestry]"]' => ['checked' => TRUE],
         ],
       ],
     ];
@@ -901,10 +929,12 @@ class IntakeForm extends FormBase {
       'intake_property_zip' => $saved_values['property']['info']['zip'],
       'intake_property_parcel_gps' => $saved_values['property']['info']['parcel_gps'],
       'intake_property_use' => $intake_property_use,
-      'intake_property_use_grazing_ac' => $saved_values['property']['land_use']['grazing_acreage'],
+      'intake_property_use_rangeland_ac' => $saved_values['property']['land_use']['rangeland_acreage'],
+      'intake_property_use_pasture_ac' => $saved_values['property']['land_use']['pasture_acreage'],
       'intake_property_use_vineyard_ac' => $saved_values['property']['land_use']['vineyards_acreage'],
       'intake_property_use_orchard_ac' => $saved_values['property']['land_use']['orchards_acreage'],
       'intake_property_use_rowcrop_ac' => $saved_values['property']['land_use']['rowcrops_acreage'],
+      'intake_property_use_forestry_ac' => $saved_values['property']['land_use']['forestry_acreage'],
       'intake_property_use_natural_ac' => $saved_values['property']['land_use']['natural_acreage'],
       'intake_property_use_other' => $saved_values['property']['land_use']['other'],
       'intake_property_use_other_ac' => $saved_values['property']['land_use']['other_acreage'],